requirements.txt
pyserial==3.4.0
folium==0.7.0
PyInstaller==3.4